requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
langchain-openai
openai==2.24.0
pinecone==8.1.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.27.1
python-dotenv==1.2.1
ragas==0.4.3
sniffio==1.3.1
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0

datasets>=2.14.0                 # RAGAS evaluation

# ── AWS Bedrock pipeline ──────────────────────────────────────────────────
boto3>=1.34.0
botocore>=1.34.0

# ── OpenSearch Serverless vector index creation (optional but recommended)
# pip install opensearch-py requests-aws4auth
opensearch-py>=2.4.0
requests-aws4auth>=1.2.3

# ── Streamlit UI ────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Voice RAG Server (Twilio + FastAPI + Qdrant) ─────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
twilio>=9.0.0
qdrant-client>=1.9.0
fastembed>=0.3.0
python-multipart>=0.0.9

